extract_country_and_city/requirements.txt
geograpy3
pandas
nltk